chvjbn

+++ mobile/funciones.php
@@ -684,7 +684,8 @@ function historialViajes($user){
 */
 function dropViaje($us, $id){
   $db = new Conexion();
-  $delete = $db->delete("DELETE FROM reservaciones WHERE idUsuario = '$us' AND idViaje = '$id'");
+  $delete = $db->delete("DELETE FROM reservaciones WHERE idUsuario = '$us' AND idViaje = '$id' AND Status != 'Confirmado'
+                          OR idUsuario = '$us' AND idViaje = '$id' AND Status != 'En curso'");
   if($delete){
     return true;
   }else{
